VideoLAN.VLC version 3.0.18 (#90601)
* VideoLAN.VLC version 3.0.18

* Update manifests/v/VideoLAN/VLC/3.0.18/VideoLAN.VLC.installer.yaml

When I created the manifest, I used https://download..., but I got a 404. Well, tried it again and now it's working... Here we go...

Co-authored-by: Muhammad Danish <mdanishkhdev@gmail.com>

* Update manifests/v/VideoLAN/VLC/3.0.18/VideoLAN.VLC.installer.yaml

When I created the manifest, I used https://download..., but I got a 404. Well, tried it again and now it's working... Here we go...

Co-authored-by: Muhammad Danish <mdanishkhdev@gmail.com>

* Update manifests/v/VideoLAN/VLC/3.0.18/VideoLAN.VLC.installer.yaml

When I created the manifest, I used https://download..., but I got a 404. Well, tried it again and now it's working... Here we go...

Co-authored-by: Muhammad Danish <mdanishkhdev@gmail.com>

* Update manifests/v/VideoLAN/VLC/3.0.18/VideoLAN.VLC.installer.yaml

When I created the manifest, I used https://download..., but I got a 404. Well, tried it again and now it's working... Here we go...

Co-authored-by: Muhammad Danish <mdanishkhdev@gmail.com>

* Update manifests/v/VideoLAN/VLC/3.0.18/VideoLAN.VLC.installer.yaml

Co-authored-by: Muhammad Danish <mdanishkhdev@gmail.com>

* Update manifests/v/VideoLAN/VLC/3.0.18/VideoLAN.VLC.installer.yaml

Co-authored-by: Muhammad Danish <mdanishkhdev@gmail.com>

* Update manifests/v/VideoLAN/VLC/3.0.18/VideoLAN.VLC.installer.yaml

Ahh, allright, I "reused" this from the previous manifest and simply adjusted the version number...

Co-authored-by: Muhammad Danish <mdanishkhdev@gmail.com>

* Update manifests/v/VideoLAN/VLC/3.0.18/VideoLAN.VLC.installer.yaml

Co-authored-by: Muhammad Danish <mdanishkhdev@gmail.com>

* Update manifests/v/VideoLAN/VLC/3.0.18/VideoLAN.VLC.installer.yaml

Co-authored-by: Muhammad Danish <mdanishkhdev@gmail.com>

* Update manifests/v/VideoLAN/VLC/3.0.18/VideoLAN.VLC.installer.yaml

Ahh, allright, I "reused" this from the previous manifest and simply adjusted the version number...

Co-authored-by: Muhammad Danish <mdanishkhdev@gmail.com>

Co-authored-by: Muhammad Danish <mdanishkhdev@gmail.com>